+++ feedbot/criteria.go
@@ -2,98 +2,7 @@ package feedbot
 
 import "github.com/mikeydub/go-gallery/service/persist"
 
-type Criteria struct {
-	EventIsValid                         Criterion
-	EventNoMoreRecentEvents              Criterion
-	IsUserCreatedEvent                   Criterion
-	IsUserFollowedEvent                  Criterion
-	IsNftCollectorsNoteAddedEvent        Criterion
-	IsCollectionCreatedEvent             Criterion
-	IsCollectionCollectorsNoteAddedEvent Criterion
-	IsCollectionTokensAddedEvent         Criterion
-	UserHasUsername                      Criterion
-	UserNoEventsBefore                   Criterion
-	FollowedUserHasUsername              Criterion
-	NftHasCollectorsNote                 Criterion
-	NftBelongsToCollection               Criterion
-	CollectionHasNfts                    Criterion
-	CollectionHasCollectorsNote          Criterion
-	CollectionHasNewTokensAdded          Criterion
-}
-
-type Criterion struct {
-	name string
-	eval func(q Query) bool
-}
-
-func newCriteria() Criteria {
-	return Criteria{
-		EventIsValid: Criterion{
-			name: "EventIsValid",
-			eval: eventIsValid,
-		},
-		EventNoMoreRecentEvents: Criterion{
-			name: "EventNoMoreRecentEvents",
-			eval: eventNoMoreRecentEvents,
-		},
-		IsUserCreatedEvent: Criterion{
-			name: "IsUserCreatedEvent",
-			eval: isUserCreatedEvent,
-		},
-		IsUserFollowedEvent: Criterion{
-			name: "IsUserFollowedEvent",
-			eval: isUserFollowedEvent,
-		},
-		IsNftCollectorsNoteAddedEvent: Criterion{
-			name: "IsNftCollectorsNoteAddedEvent",
-			eval: isNftCollectorsNoteAddedEvent,
-		},
-		IsCollectionCreatedEvent: Criterion{
-			name: "IsCollectionCreatedEvent",
-			eval: isCollectionCreatedEvent,
-		},
-		IsCollectionCollectorsNoteAddedEvent: Criterion{
-			name: "IsCollectionCollectorsNoteAddedEvent",
-			eval: isCollectionCollectorsNoteAddedEvent,
-		},
-		IsCollectionTokensAddedEvent: Criterion{
-			name: "IsCollectionTokensAddedEvent",
-			eval: isCollectionTokensAddedEvent,
-		},
-		UserHasUsername: Criterion{
-			name: "UserHasUsername",
-			eval: userHasUsername,
-		},
-		UserNoEventsBefore: Criterion{
-			name: "UserNoEventsBefore",
-			eval: userNoEventsBefore,
-		},
-		FollowedUserHasUsername: Criterion{
-			name: "FollowedUserHasUsername",
-			eval: followedUserHasUsername,
-		},
-		NftHasCollectorsNote: Criterion{
-			name: "NftHasCollectorsNote",
-			eval: nftHasCollectorsNote,
-		},
-		NftBelongsToCollection: Criterion{
-			name: "NftBelongsToCollection",
-			eval: nftBelongsToCollection,
-		},
-		CollectionHasNfts: Criterion{
-			name: "CollectionHasNfts",
-			eval: collectionHasNfts,
-		},
-		CollectionHasCollectorsNote: Criterion{
-			name: "CollectionHasCollectorsNote",
-			eval: collectionHasCollectorsNote,
-		},
-		CollectionHasNewTokensAdded: Criterion{
-			name: "CollectionHasNewTokensAdded",
-			eval: collectionHasNewTokensAdded,
-		},
-	}
-}
+var baseCriteria = []func(Query) bool{eventIsValid, eventNoMoreRecentEvents, userHasUsername}
 
 func eventIsValid(q Query) bool {
 	if q.EventID == "" {
